t204: session-scoped polling with exponential backoff and visibility throttling
Implements t203 prerequisite (extract jobSlice from sessionsSlice) and
t204 (session-scoped polling with backoff + visibility manager).

NEW: src/store/slices/jobSlice.js — owns job state (currentJobId,
liveToolCalls, sessionJobs, pendingConfirmation, pendingActionCard) and
pollJob(jobId, sessionId) with:
- Session-scoped: no currentJobId check, multiple sessions poll independently
- Exponential backoff: 1s → 5s after 10 polls → 10s cap after 30 polls
- Progress-detected reset: backoff resets to 1s on tool_calls length increase
- Visibility throttling: slows to 15s when document.hidden, immediate poll
  on tab-restore via visibility-manager subscription

NEW: src/utils/visibility-manager.js — singleton document.visibilitychange
listener with subscribe/unsubscribe API (onVisibilityChange).

EDIT: src/store/slices/sessionsSlice.js — remove extracted job state,
actions, selectors, and reducer cases. Update pollJob call sites in
streamMessage, confirmToolCall, and rejectToolCall to pass sessionId.

EDIT: src/store/index.js — import and wire jobSlice into combined
DEFAULT_STATE, actions, selectors, and reducer chain.

Resolves #1032
For #1031
